requirements.txt
sqlalchemy
sqlalchemy-utils
requests
psycopg2
schedule
python-dotenv
